Move code to avoid same named file
