+++ plugins/core/src/main/kotlin/de/fayard/refreshVersions/core/internal/versions/VersionsPropertiesWriting.kt
@@ -1,9 +1,59 @@
 package de.fayard.refreshVersions.core.internal.versions
 
+import de.fayard.refreshVersions.core.Version
+import de.fayard.refreshVersions.core.extensions.gradle.toModuleIdentifier
+import de.fayard.refreshVersions.core.internal.DependencyWithVersionCandidates
+import de.fayard.refreshVersions.core.internal.RefreshVersionsConfigHolder
+import de.fayard.refreshVersions.core.internal.getVersionPropertyName
+import de.fayard.refreshVersions.core.internal.isAVersionAlias
 import de.fayard.refreshVersions.core.internal.versions.VersionsPropertiesModel.Companion.availableComment
 import de.fayard.refreshVersions.core.internal.versions.VersionsPropertiesModel.Section.Comment
 import de.fayard.refreshVersions.core.internal.versions.VersionsPropertiesModel.Section.VersionEntry
+import org.gradle.api.Project
 import java.io.File
+
+internal fun Project.updateVersionsProperties(
+    dependenciesWithLastVersion: List<DependencyWithVersionCandidates>
+) {
+    val versionKeyReader = RefreshVersionsConfigHolder.versionKeyReader
+    val file = RefreshVersionsConfigHolder.versionsPropertiesFile
+
+    val model = VersionsPropertiesModel.readFrom(file)
+    val updates = dependenciesWithLastVersion.associateBy { (moduleId, _, _) ->
+        getVersionPropertyName(moduleId.toModuleIdentifier(), versionKeyReader)
+    }
+
+    val newModel = model.copy(
+        sections = model.sections.map { section ->
+            when (section) {
+                is Comment -> section
+                is VersionEntry -> {
+                    if (section.currentVersion.isAVersionAlias()) return@map section
+                    val update = updates[section.key] ?: return@map section
+                    section.copy(
+                        availableUpdates = update.versionsCandidates.map { it.value }
+                    )
+                }
+            }
+        }
+    )
+    newModel.writeTo(file)
+}
+
+internal fun writeWithAddedVersions(
+    versionsFile: File,
+    propertyName: String,
+    versionsCandidates: List<Version>
+) {
+    val model = VersionsPropertiesModel.readFrom(versionsFile)
+
+    val newModel = model + VersionEntry(
+        key = propertyName,
+        currentVersion = versionsCandidates.first().value,
+        availableUpdates = versionsCandidates.drop(1).map { it.value }
+    )
+    newModel.writeTo(versionsFile)
+}
 
 internal fun VersionsPropertiesModel.writeTo(versionsPropertiesFile: File) {
     val finalModel = this.copy(
